requirements.txt
aiosmtpd==1.4.6
annotated-types==0.7.0
asgiref==3.9.0
atpublic==6.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.7.9
charset-normalizer==3.4.2
crispy-bootstrap5==2025.6
Django==4.2.23
django-bootstrap-datepicker-plus==5.0.5
django-bootstrap-v5==1.0.11
django-crispy-forms==2.4
django-resized==1.0.3
django-tinymce==4.1.0
fontawesomefree==6.6.0
idna==3.10
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
